refactor works need fix logging
